move function from script to background task
